+++ CHANGELOG.md
@@ -72,6 +72,7 @@ The next release will require at least [Go 1.24].
 
 - Fix `go.opentelemetry.io/otel/exporters/prometheus` to deduplicate suffixes if already present in metric name when UTF8 is enabled. (#7088)
 - Fix the `go.opentelemetry.io/otel/exporters/stdout/stdouttrace` self-observability component type and name. (#7195)
+- Fix partial export count metric in `go.opentelemetry.io/otel/exporters/stdout/stdouttrace`. (#7199)
 
 <!-- Released section -->
 <!-- Don't change this section unless doing release -->